pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "play-helloworld-auth-api"
authors = ["dlake82 <dlake0828@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
djangorestframework = "^3.14.0"
pygments = "^2.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
